upload new repo

+++ transformers/tests/trainer/test_data_collator.py
@@ -1052,9 +1052,7 @@ class TFDataCollatorIntegrationTest(unittest.TestCase):
 
         # confirm that every token is either the original token or [MASK]
         self.assertTrue(
-            tf.reduce_all(
-                (batch["input_ids"] == tf.cast(inputs, tf.int64)) | (batch["input_ids"] == tokenizer.mask_token_id)
-            )
+            tf.reduce_all((batch["input_ids"] == inputs) | (batch["input_ids"] == tokenizer.mask_token_id))
         )
 
         # numpy call
